requirements.txt
feedparser
fpdf
requests
python-dotenv
